+++ index.js
@@ -1,6 +1,15 @@
 const {controlBlock} = require('./scripts');
 const {hashForTree} = require('./scripts');
 const {leafHash} = require('./scripts');
+const {signHash} = require('./keys');
 const {tweakHash} = require('./keys');
+const {v1OutputScript} = require('./scripts');
 
-module.exports = {controlBlock, hashForTree, leafHash, tweakHash};
+module.exports = {
+  controlBlock,
+  hashForTree,
+  leafHash,
+  signHash,
+  tweakHash,
+  v1OutputScript,
+};