+++ scripts/index.js
@@ -1,5 +1,6 @@
 const controlBlock = require('./control_block');
 const hashForTree = require('./hash_for_tree');
 const leafHash = require('./leaf_hash');
+const v1OutputScript = require('./v1_output_script');
 
-module.exports = {controlBlock, hashForTree, leafHash};
+module.exports = {controlBlock, hashForTree, leafHash, v1OutputScript};